requirements.txt
Django==2.1.5
django-rest-auth==0.9.3
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.5
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
six==1.12.0